Create dockerfile and deploy script (#2)
Dockerize project for quick deployments. Added a quick Dockerfile to build the NGINX image and copy the contents of the dist/ directory to the directory served by NGINX in the container.

I also made a quick deploy script, that runs the necessary node scripts, then builds and starts the Docker container and binds port 80 of the Docker container to port 80 of the host machine.